+++ library/src/main/java/com/zen/android/eroid/injection/component/IDroidComponent.java
@@ -1,4 +1,4 @@
-package com.zen.android.eroid.inject.component;
+package com.zen.android.eroid.injection.component;
 
 import com.zen.android.eroid.ui.DroidActivity;
 import com.zen.android.eroid.ui.DroidFragment;
